moves pulser outside of experiment folder
